feat: route DPS reports through dedupe relay; fix capture stall on long sessions
DemonSoul - ZDPS Meter v0.28.0

This release moves the end-of-fight DPS report off the old direct Discord
webhook and onto the DemonSoul server relay, and fixes a packet-capture bug
that could make the meter — and especially the chat logs — quietly stop
updating after a long session.

What's new
- DPS reports now go through the server relay. When several party members
  run the meter, their duplicate reports for the same fight are collapsed
  into a single Discord post instead of spamming the channel.
- The report image is now rendered cleanly on the server from the fight
  data. No more screenshots — the graph always shows the DPS table (not
  whatever tab you happened to be on) and is never the dark, half-transparent
  capture it used to be.

Fixes
- Fixed a TCP sequence-number wraparound bug that could permanently freeze
  packet reassembly. The symptom was combat stats and chat logs silently
  stopping after about an hour, or after the game sat idle for a while and
  then activity resumed. Capture now handles the 32-bit wrap correctly and
  self-heals instead of wedging.

+++ src-tauri/src/packets/utils.rs
@@ -52,8 +52,13 @@ fn ip_to_str(ip: &[u8; 4]) -> String {
 }
 
 pub struct TCPReassembler {
-    pub cache: BTreeMap<usize, Vec<u8>>, // sequence -> payload
-    pub next_seq: Option<usize>,         // next expected sequence
+    // Keyed by the raw 32-bit TCP sequence number. All sequence arithmetic must
+    // be done mod 2^32 (these are u32, not usize) so it survives the wraparound
+    // that happens once a connection's sequence counter crosses 2^32 — storing
+    // them as usize and using usize::wrapping_add wraps at 2^64 instead, which
+    // permanently wedged the stream after a while (random ISN => random time).
+    pub cache: BTreeMap<u32, Vec<u8>>, // sequence -> payload
+    pub next_seq: Option<u32>,         // next expected sequence
     pub _data: Vec<u8>,
 }
 
@@ -66,7 +71,7 @@ impl TCPReassembler {
         }
     }
 
-    pub fn clear_reassembler(&mut self, seq_number: usize) {
+    pub fn clear_reassembler(&mut self, seq_number: u32) {
         self.cache = BTreeMap::new();
         self._data.clear();
         self.next_seq = Some(seq_number);
